fix/#25-move-enum-TrainType
